Multi-digit classifier/requirements.txt
numpy~=1.25.2
matplotlib~=3.7.2
torchvision~=0.15.2
pillow~=10.0.0
tensorboard~=2.14.0
